fix: Allow for negative XAXISTILT values in imod .com files (#10)
Dear developers, 

Thanks for this useful package to convert between the different
alignment file types!

We ran into an issue that when we read an IMOD .com file that had a
negative XAXISTILT, it would not be grabbed by the regex.

This PR solves that (allows for 0 or 1 negative sign) and makes sure
this behavior is tested